src/field.py
class Field:
    """
    Represents a rectangular field for the car to move within.
    """

    def __init__(self, width, height):
        """
        Initializes the field with the given width and height.
        Raises ValueError if width or height is negative or not an integer.
        """
        if not isinstance(width, int) or not isinstance(height, int):
            raise ValueError("Width and height must be integers.")
        if width < 0 or height < 0:
            raise ValueError("Width and height must be non-negative.")
        
        self.width = width
        self.height = height
        self.occupied_positions = {}  # Dictionary to keep track of occupied positions by car identifiers

    def is_within_bounds(self, x, y):
        """
        Checks if the given (x, y) position is within the field boundaries.
        """
        return 0 <= x < self.width and 0 <= y < self.height

    def simulate_multiple_cars(self, cars_with_commands):
        """
        Simulates the movement of multiple cars and checks for collisions.
        
        :param cars_with_commands: A list of tuples, each containing a Car instance and a string of commands.
        :return: Formatted string indicating collision or "no collision".
        """
        max_steps = max(len(commands) for _, commands in cars_with_commands)
        
        for step in range(max_steps):
            for car, commands in cars_with_commands:
                if step < len(commands):
                    command = commands[step]
                    if command == 'L':
                        car.rotate_left()
                    elif command == 'R':
                        car.rotate_right()
                    elif command == 'F':
                        potential_x, potential_y = car.x, car.y

                        if car.direction == 'N':
                            potential_y += 1
                        elif car.direction == 'E':
                            potential_x += 1
                        elif car.direction == 'S':
                            potential_y -= 1
                        elif car.direction == 'W':
                            potential_x -= 1

                        # Check if the move is within bounds
                        if self.is_within_bounds(potential_x, potential_y):
                            # Check if the position is already occupied
                            for other_car_id, position in self.occupied_positions.items():
                                if position == (potential_x, potential_y):
                                    # Sort car identifiers alphabetically before returning
                                    car_ids = sorted([car.identifier, other_car_id])
                                    return f"{car_ids[0]} {car_ids[1]}\n{potential_x} {potential_y}\n{step + 1}"
                            
                            # Update car position
                            self.occupied_positions.pop(car.identifier, None)
                            car.x, car.y = potential_x, potential_y
                            self.occupied_positions[car.identifier] = (car.x, car.y)

        return "no collision"